raycasting functional alex July 28 2022

+++ srcs/utils/draw_line.c
@@ -6,13 +6,13 @@
 /*   By: jvalenci <jvalenci@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/16 10:32:44 by jvalenci          #+#    #+#             */
-/*   Updated: 2022/07/26 17:33:37 by jvalenci         ###   ########.fr       */
+/*   Updated: 2022/07/28 10:51:22 by jvalenci         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include"../../includes/cub3d.h"
 
-void    plot_line_low(t_map *m, int *xy0, int *xy1)
+void    plot_line_low(t_map *m, float *xy0, float *xy1)
 {
     m->l->dx = xy1[0] - xy0[0];
     m->l->dy = xy1[1] - xy0[1];
@@ -39,7 +39,7 @@ void    plot_line_low(t_map *m, int *xy0, int *xy1)
     }
 }
 
-void    plot_line_hight(t_map *m, int *xy0, int *xy1)
+void    plot_line_hight(t_map *m, float *xy0, float *xy1)
 {
     m->l->dx = xy1[0] - xy0[0];
     m->l->dy = xy1[1] - xy0[1];
@@ -76,14 +76,14 @@ https://en.wikipedia.org/wiki/Bresenham%27s_line_algorithm
 // m->l->pdxy[0] = cos(angle_r) * 100 + x0 ;
 // m->l->pdxy[1] = sin(angle_r) * 100 + y0 ;
 */
-void plot_line(t_map *m, int x0, int y0)
+void plot_line(t_map *m, float x0, float y0)
 {
-    int xy0[2];
+    float xy0[2];
     // int xy1[2];
 
     xy0[0] = x0;
     xy0[1] = y0;
-    if (abs(m->l->pdxy[1] - xy0[1]) < abs(m->l->pdxy[0] - xy0[0]))
+    if (abs((int)m->l->pdxy[1] - (int)xy0[1]) < abs((int)m->l->pdxy[0] - (int)xy0[0]))
     {
         if (xy0[0] > m->l->pdxy[0])
             plot_line_low(m, m->l->pdxy, xy0);
